requirements.txt
# Streamlit is still needed for the Gmail Agent page, will be removed when fully ported to React
streamlit==1.44.1
langchain==0.3.23
langchain-groq==0.3.2
langchain-core==0.3.51
langchain-community==0.3.21
python-dotenv==1.1.0
PyMuPDF==1.25.5
newspaper4k==0.9.3.1
lxml_html_clean
openai==1.72.0
langchain-openai==0.3.12
pinecone==6.0.2
langchain-pinecone==0.2.5
youtube-transcript-api==1.0.3
google-generativeai==0.8.4
numpy==2.2.4
pandas==2.2.3
faiss-cpu==1.10.0
google-api-python-client==2.166.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pytube2==15.0.12
yt-dlp==2025.3.31
Pillow==11.1.0
